+++ src/model.rs
@@ -34,6 +34,73 @@ pub struct Item {
     pub rarity: Rarity,
 }
 
+impl Item {
+    pub fn element(self) -> &'static str {
+        crate::simulation::item_element(self.name)
+    }
+    pub fn stats(self) -> Stats {
+        crate::simulation::item_stats(self)
+    }
+}
+
+#[derive(Clone, Copy, Debug)]
+pub struct Stats {
+    pub crit_dmg: u16,
+    pub crit_rate: u16,
+    pub atk: u16,
+    pub def: u16,
+    pub spd: u16,
+    pub elemental_atk: u16,
+    pub hp: u16,
+    pub poise: u16,
+}
+
+impl Stats {
+    pub const fn character(
+        rarity: Rarity,
+        atk: u16,
+        def: u16,
+        spd: u16,
+        hp: u16,
+        poise: u16,
+    ) -> Self {
+        Self {
+            crit_dmg: 150,
+            crit_rate: if matches!(rarity, Rarity::Five) {
+                10
+            } else {
+                5
+            },
+            atk,
+            def,
+            spd,
+            elemental_atk: if matches!(rarity, Rarity::Five) {
+                115
+            } else {
+                100
+            },
+            hp,
+            poise,
+        }
+    }
+    pub const fn weapon(rarity: Rarity, atk: u16, crit_rate: u16, crit_dmg: u16) -> Self {
+        Self {
+            crit_dmg,
+            crit_rate,
+            atk,
+            def: 0,
+            spd: 0,
+            elemental_atk: if matches!(rarity, Rarity::Five) {
+                18
+            } else {
+                12
+            },
+            hp: 0,
+            poise: 0,
+        }
+    }
+}
+
 #[derive(Clone, Debug)]
 pub struct WishResult {
     pub item: Item,
@@ -48,17 +115,28 @@ pub enum Banner {
     Astraea,
     Kaelis,
     Seraphine,
+    Vaughn,
+    Steven,
     Weapon,
 }
 
 impl Banner {
-    pub const ALL: [Self; 4] = [Self::Astraea, Self::Kaelis, Self::Seraphine, Self::Weapon];
+    pub const ALL: [Self; 6] = [
+        Self::Astraea,
+        Self::Kaelis,
+        Self::Seraphine,
+        Self::Vaughn,
+        Self::Steven,
+        Self::Weapon,
+    ];
 
     pub const fn title(self) -> &'static str {
         match self {
             Self::Astraea => "Starfall Archive",
             Self::Kaelis => "Embers of Rebellion",
             Self::Seraphine => "Whispers in Bloom",
+            Self::Vaughn => "Violet Oath Eternal",
+            Self::Steven => "Veilfire Covenant",
             Self::Weapon => "Incarnate Armaments",
         }
     }